pyproject.toml
[project]
name = "discord-issues"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
]